Peer Review
The section around lines 200 is the only thing that I had input on, and
there wasn't anything I could really find wrong with it, but I put some
comments in that section. I tried to see if I could fix your problem
with changing the direction in which the colors of the circles changes
but I couldn't do it. Other than that, I couldn't really find any errors
in your code or something you could improve upon.